requirements-win.txt
altgraph==0.17.3
lxml==4.9.2
numpy==1.24.2
pandas==1.5.3
pefile==2023.2.7
Pillow==9.4.0
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.7.1
pywin32-ctypes==0.2.0
PyYAML==6.0
six==1.16.0
XlsxWriter==3.0.8
